+++ src/index.js
@@ -5,8 +5,8 @@ import App from "./App";
 import "./index.css";
 
 render(
-	<React.StrictMode>
-		<App />
-	</React.StrictMode>,
-	document.getElementById("root")
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+  document.getElementById("root")
 );